requirements.txt
aiofiles==25.1.0
anyio==4.12.1
blinker==1.9.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cryptography==46.0.4
dash==3.4.0
deprecation==2.1.0
dotenv==0.9.9
Flask==3.1.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
importlib_metadata==8.7.1
itsdangerous==2.2.0
Jinja2==3.1.6
jwcrypto==1.5.6
MarkupSafe==3.0.3
narwhals==2.15.0
nest-asyncio==1.6.0
packaging==26.0
plotly==6.5.2
pycparser==3.0
PyJWT==2.10.1
python-dotenv==1.2.1
python-keycloak==7.0.3
requests==2.32.5
requests-toolbelt==1.0.0
retrying==1.4.2
setuptools==80.10.2
typing_extensions==4.15.0
urllib3==2.6.3
Werkzeug==3.1.5
zipp==3.23.0
